defender

+++ database/migrations/2015_02_23_161103_create_defender_role_user_table.php
@@ -11,10 +11,10 @@ class CreateDefenderRoleUserTable extends Migration
     public function up()
     {
         Schema::create(config('defender.role_user_table', 'role_user'), function (Blueprint $table) {
-            $table->unsignedInteger('user_id')->index();
+            $table->unsignedBigInteger('user_id')->index();
             $table->foreign('user_id')->references('id')->on(config('auth.table', 'users'))->onDelete('cascade');
 
-            $table->unsignedInteger(config('defender.role_key', 'role_id'))->index();
+            $table->unsignedBigInteger(config('defender.role_key', 'role_id'))->index();
             $table->foreign(config('defender.role_key', 'role_id'))->references('id')
                   ->on(config('defender.role_table', 'roles'))
                   ->onDelete('cascade');
