+++ database/migrations/2015_02_23_161105_create_defender_permission_role_table.php
@@ -13,12 +13,12 @@ class CreateDefenderPermissionRoleTable extends Migration
     public function up()
     {
         Schema::create(config('defender.permission_role_table', 'permission_role'), function (Blueprint $table) {
-            $table->unsignedInteger(config('defender.permission_key', 'permission_id'))->index();
+            $table->unsignedBigInteger(config('defender.permission_key', 'permission_id'))->index();
             $table->foreign(config('defender.permission_key', 'permission_id'))->references('id')
                   ->on(config('defender.permission_table', 'permissions'))
                   ->onDelete('cascade');
 
-            $table->unsignedInteger(config('defender.role_key', 'role_id'))->index();
+            $table->unsignedBigInteger(config('defender.role_key', 'role_id'))->index();
             $table->foreign(config('defender.role_key', 'role_id'))->references('id')
                   ->on(config('defender.role_table', 'roles'))
                   ->onDelete('cascade');
